chore: remove unused files, dependencies and rename to .jsx
