+++ app/entry.server.jsx
@@ -1,0 +1,1 @@
+export { handleRequest as default } from '@netlify/remix-adapter';
