--- app/entry.server.tsx
@@ -1,1 +1,0 @@
-export { handleRequest as default } from "@netlify/remix-adapter";
