--- globals.d.ts
@@ -1,10 +1,0 @@
-export {};
-declare global {
-  interface ProcessEnv {
-    NODE_ENV: "development" | "production" | "test";
-  }
-  interface Process {
-    env: ProcessEnv;
-  }
-  let process: Process;
-}
